refactor: for use inside the labxchange app
* uses labxchange learning contexts where available, instead of openedx
* adds a simple AppConfig with no dependencies on open edx,
  so the plugin can behave like a normal app within LabXchange
